review: fix test and schema

+++ database/models/product.model.js
@@ -28,12 +28,12 @@ const Product = (sequelize) => {
       defaultValue: 5
     },
     buyPrice: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.INTEGER,
       allowNull: false,
       defaultValue: 0
     },
     salePrice: {
-      type: Sequelize.FLOAT,
+      type: Sequelize.INTEGER,
       allowNull: false,
       defaultValue: 0
     }
